Another humongus commit
yeah yeah I know:"like wtf man, commit smaller and more often!"
I promise in the future.

There is a core now that can run with a couple of chaiscript
commands that can be run to make things happen in the game engine

You can initialize some modules and create a grid.
The grid isn't actually drawn yet, but it's there!