GGE Running, porting to guile completed
perl scripts expanded greated but will need still more work due
 to special cases. On pause for now.
Removed some gge_ commands as command generation is slightly altered
Command class for graphics added.
Clearing screen, drawing and rendering now possible in guile
None is an actual GGE class so that it is not a nullptr. Does, as one
 might suspect, nothing.
None is always instatiated.
nullptr means no gge module exists. This is so that if a module tries to
work with a module not instantiated it will crash.